Move all animation-related logic to SheetAnimationCoordinator
